requirements.txt
pillow==10.3.0
streamlit==1.36.0
